Swift backend working
The pure swift symbolic math backend is working. That being said, it all gets nasty because it don't simplify anything. I need to look into how that is done. Looking at the profiler, the next thing to optimize is the eval (no surprise). I may be able to alieve some of that by inlining the functions, but otherwise, I think it's partially due to the complexity of the equation trees.